Add --uninstall option
https://github.com/shirok/get-gauche/issues

Currently, this takes the same step as installation to determine the version,
download the tarball, configure, and then simply executes "make uninstall".
We do this because we don't know if previously untarred source tree is still
available.

Ideally, it would be much simpler if Gauche installation generates "uninstall"
script and install it altogether.